chore: refactor file names
