+++ src/telemetry/WatchdogClient.ts
