+++ src/telemetry/metricUtils.ts
