+++ src/telemetry/watchdog/ClearcutSender.ts
